requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
pylibmc==1.4.2
django-heroku-memcacheify==0.8
-r requirements/production.txt